Final bits of refactor - project structure
