+++ src/components/atom/ErrorWrapper/ErrorWrapper.spec.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+import Component from '.';
+
+describe('ErrorWrapper component', () => {
+  it('renders correctly', () => {
+    const component = <Component />;
+
+    const wrapper = render(component);
+    
+    expect(wrapper).toMatchSnapshot();
+  });
+});
