+++ src/components/atom/FormWrapper/index.jsx
@@ -1,6 +1,16 @@
 import styled from 'styled-components';
+import theme from '../../../global/theme';
 
 export default styled.form`
-  margin: 16px 24px;
-  height: 100%;
+justify-content: center;
+width: 100%;
+height: 100%;
+background-color: white;
+@media ${theme.media.tablet}{
+  margin: 25px 30px;
+  height: min-content;
+}
+@media ${theme.media.desktop}{
+  margin: 45px 70px;
+}
 `;
