+++ src/components/atom/ProjectSummary/TextWrapper/TextWrapper.spec.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+import Component from '.';
+
+describe('TextWrapper component', () => {
+  it('renders correctly', () => {
+    const component = <Component />;
+
+    const wrapper = render(component);
+    
+    expect(wrapper).toMatchSnapshot();
+  });
+});
