+++ src/components/atom/ProjectSummary/TextWrapper/index.jsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export default styled.div`
+  align-self: flex-end;
+  position: relative;
+  width: 100%;
+`;
