+++ src/components/atom/ProjectSummary/index.js
@@ -1,0 +1,11 @@
+import ClientName from './ClientName';
+import ProjectName from './ProjectName';
+import SummaryBox from './SummaryBox';
+import TextWrapper from './TextWrapper';
+
+export {
+  ClientName,
+  ProjectName,
+  SummaryBox,
+  TextWrapper
+}
