+++ src/components/organism/ProjectSummary/index.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { SummaryBox, ProjectName, ClientName } from './styles';
-import { TextWrapper } from '../../../global/styles';
+import { SummaryBox, ProjectName, ClientName, TextWrapper } from '../../atom/ProjectSummary';
 
 const ProjectSummary = ({ projectName, projectClient, projectImageURL, onClick }) => {
   return (
