+++ src/global/helpers/index.js
@@ -1,0 +1,7 @@
+import { onInputChange } from './inputs';
+import { shuffleAndSliceArray } from './arrays';
+
+export {
+  onInputChange,
+  shuffleAndSliceArray
+}
