+++ src/global/helpers/inputs/index.js
@@ -1,0 +1,2 @@
+/* eslint-disable import/prefer-default-export */
+export const onInputChange = set => event => set(event.target.value);
